requirements-macappbundle.txt
pyinstaller
std-nslog
Pillow
